# lib/api-zod/src/generated/types/backlinkInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LaunchCore SEO Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface BacklinkInput {
  domain: string;
  /** @nullable */
  crawlId?: number | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  niche?: string | null;
}