# lib/api-zod/src/generated/types/backlinkJob.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LaunchCore SEO Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { BacklinkJobStatus } from './backlinkJobStatus';

export interface BacklinkJob {
  id: number;
  domain: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  niche?: string | null;
  status: BacklinkJobStatus;
  /** @nullable */
  opportunitiesFound?: number | null;
  createdAt: Date;
  /** @nullable */
  completedAt?: Date | null;
}