# lib/api-zod/src/generated/types/crawlSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LaunchCore SEO Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { Crawl } from './crawl';
import type { CrawlSummaryIssueCounts } from './crawlSummaryIssueCounts';
import type { CrawlSummaryTodoCounts } from './crawlSummaryTodoCounts';
import type { SeoIssue } from './seoIssue';

export interface CrawlSummary {
  crawl: Crawl;
  issueCounts: CrawlSummaryIssueCounts;
  todoCounts: CrawlSummaryTodoCounts;
  topIssues: SeoIssue[];
}